adding some styles font-awesome, swal, bootstrap, and restructuring the components

+++ smart-currencyer-client/src/app/components/currencies/currencies.component.ts
@@ -1,6 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { CurrencyService } from './../services/currency.service';
-import { Currency } from './../classes/currency';
+import { CurrencyService } from './../../services/currency.service';
+import { Currency } from './../../classes/currency';
+
+import swal from 'sweetalert2';
+declare var jquery:any;
+declare var $ :any;
 
 
 @Component({
@@ -29,13 +33,36 @@ export class CurrenciesComponent implements OnInit {
     if (!symbol) { return; }
     this.currencyService.addCurrency({ name, symbol } as Currency)
       .subscribe(currency => {
+        $('#currencyModal').modal('hide');
+
+        swal({
+          title: 'Currency created',
+          type: 'success',
+        })
+
         this.currencies.push(currency);
       });
   }
 
   delete(currency: Currency): void {
-    this.currencies = this.currencies.filter(c => c !== currency);
-    this.currencyService.deleteCurrency(currency).subscribe();
+    swal({
+      title: 'Are you sure?',
+      text: `All wallets with ${currency.name} currency will be destroy`,
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Delete',
+      cancelButtonText: 'Cancel'
+    }).then((result) => {
+      if (result.value) {
+        swal(
+          'Deleted!',
+          `The currency ${currency.name} was deleted`,
+          'success'
+        )
+        this.currencies = this.currencies.filter(c => c !== currency);
+        this.currencyService.deleteCurrency(currency).subscribe();
+      }
+    })
   }
